[04] find the largest palindrome
my initial try was to break out of the loop as soon as I found one, but
that did not work because of the way a double loop works. This is the
result I got

995  x  583  =  580085

However, this misses a range of factor smaller than 995 and greater than
583 which give a larger product.

This is why I added a variable Solution which checks if we have already
reached this product.

I am sure this can be improved by exiting a loop at some point (half
point?)